refactor: user registration, deletion and activation service
