+++ src/main/java/fast/bank/api/controllers/UserController.java
@@ -3,9 +3,7 @@ package fast.bank.api.controllers;
 import fast.bank.api.domain.user.dto.UserDetailingData;
 import fast.bank.api.domain.user.dto.UserRegistrationData;
 import fast.bank.api.domain.user.repository.UserRepository;
-import fast.bank.api.domain.user.service.activation.UserActivationService;
-import fast.bank.api.domain.user.service.deletion.UserDeletionService;
-import fast.bank.api.domain.user.service.registration.UserRegistrationService;
+import fast.bank.api.domain.user.service.UserService;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,48 +19,42 @@ import org.springframework.web.util.UriComponentsBuilder;
 public class UserController {
 
     @Autowired
-    private UserRepository repository;
+    private UserRepository userRepository;
 
     @Autowired
-    UserRegistrationService registrationService;
-
-    @Autowired
-    UserDeletionService deletionService;
-
-    @Autowired
-    UserActivationService activationService;
+    private UserService userService;
 
     @PostMapping
     @Transactional
     public ResponseEntity<UserDetailingData> register(@RequestBody @Valid UserRegistrationData data, UriComponentsBuilder uriBuilder) {
-        var user = registrationService.createUser(data);
+        var user = userService.register(data);
         var uri = uriBuilder.path("api/v1/users/{id}").buildAndExpand(user.getRegistry()).toUri();
         return ResponseEntity.created(uri).body(new UserDetailingData(user));
     }
 
     @GetMapping
     public ResponseEntity<Page<UserDetailingData>> list(@PageableDefault(size = 10, sort = {"registry"}) Pageable pagination) {
-        var page = repository.findAllByIsActiveTrue(pagination).map(UserDetailingData::new);
+        var page = userRepository.findAllByIsActiveTrue(pagination).map(UserDetailingData::new);
         return ResponseEntity.ok(page);
     }
 
     @GetMapping("/{registry}")
     public ResponseEntity listById(@PathVariable Long registry) {
-        var user = repository.getReferenceById(registry);
+        var user = userRepository.getReferenceById(registry);
         return ResponseEntity.ok(new UserDetailingData(user));
     }
 
     @DeleteMapping("/{registry}")
     @Transactional
     public ResponseEntity delete(@PathVariable Long registry) {
-        deletionService.logicalUserDeletion(registry);
+        userService.delete(registry);
         return ResponseEntity.noContent().build();
     }
 
     @PutMapping("/activate/{registry}")
     @Transactional
     public ResponseEntity activate(@PathVariable Long registry) {
-        var user = activationService.activate(registry);
+        var user = userService.activate(registry);
         return ResponseEntity.ok(user);
     }
 }
