+++ src/main/java/fast/bank/api/domain/user/service/validation/deletion/UserDeletionValidators.java
@@ -1,4 +1,4 @@
-package fast.bank.api.domain.user.service.deletion.validation;
+package fast.bank.api.domain.user.service.validation.deletion;
 
 public interface UserDeletionValidators {
     void validate(Long number);
