+++ src/main/java/fast/bank/api/domain/user/service/validation/registration/UniqueUser.java
@@ -1,8 +1,7 @@
-package fast.bank.api.domain.user.service.registration.validation;
+package fast.bank.api.domain.user.service.validation.registration;
 
 import fast.bank.api.domain.user.dto.UserRegistrationData;
 import fast.bank.api.domain.user.repository.UserRepository;
-import fast.bank.api.domain.user.service.registration.validation.UserRegistrationValidators;
 import fast.bank.api.infra.exception.validation.ValidException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
